requirements.txt
# Concurrency and Asynchronous Execution
gevent>=21.12.0

# File Handling
python-magic==0.4.24

# FastAPI Framework and Related Dependencies
fastapi>=0.111.0
uvicorn>=0.23.2
pydantic>=2.0

# Monitoring and Metrics
prometheus-client>=0.20.0
prometheus-fastapi-instrumentator>=6.1.0
loguru==0.7.2 

# Asynchronous File and HTTP Handling
aiofiles>=23.1.0
aiohttp>=3.9.5 
aiosignal==1.3.1

# Multipart and WebSocket Support
python-multipart>=0.0.7
websockets==15.0

# Langraph
litellm==1.52.8
langchain-community>=0.2.15
langchain-core>=0.2.5 
langchain-openai>=0.1.8  
langgraph==0.2.74
# langchain-huggingface>=0.0.3
pypdf
tiktoken

# Vector Database
langchain_postgres
# -f https://download.pytorch.org/whl/cpu
# torch==2.3.1
langchain-text-splitters>=0.2.1
# sentence-transformers>=3.0.1
# transformers==4.41.2


# LLM Tracing
langfuse==2.60.0

# Semantic Chunker Splitter
advanced-chunker

# kafka
kafka-python==2.0.2
confluent-kafka
aiokafka
psycopg[binary]==3.2.9

# Redis
redis[asyncio]

bs4
nemoguardrails
emoji